move helper function from fixes to sparsefuncs
